Refactor lc call slips w history (#699)
* Refactor NewlyCataloged to LcCallSlips

- Clearer communication between ATT team, DACS, and selectors


* Finish refactoring

* Add config change

---------
