+++ app/models/oclc/lc_call_slips/lc_call_slip_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Oclc
+  module LcCallSlips
+    class LcCallSlipJob < LibJob
+      attr_reader :report_downloader
+    end
+  end
+end
