+++ app/views/lc_call_slips_mailer/report.text.erb
