+++ config/application.rb
@@ -40,7 +40,7 @@ module IlsApps
 
     config.alma_sftp = config_for(:alma_sftp)
     config.oclc_sftp = config_for(:oclc_sftp)
-    config.newly_cataloged = config_for(:newly_cataloged)
+    config.lc_call_slips = config_for(:lc_call_slips)
     config.peoplesoft = config_for(:peoplesoft)
     config.pod = config_for(:pod)
     config.aspace = config_for(:aspace)
